+++ dramatiq/cli.py
@@ -77,6 +77,9 @@ examples:
   # Run with a broker named "broker" defined as attribute of "app" in "some_module".
   $ dramatiq some_module:app.broker
 
+  # Run with a callable that sets up a broker.
+  $ dramatiq some_module:setup_broker
+
   # Auto-reload dramatiq when files in the current directory change.
   $ dramatiq --watch . some_module
 
@@ -117,13 +120,17 @@ def import_object(value):
 
 
 def import_broker(value):
-    module, broker = import_object(value)
-    if broker is None:
+    module, broker_or_callable = import_object(value)
+    if broker_or_callable is None:
         return module, get_broker()
 
-    if not isinstance(broker, Broker):
+    if callable(broker_or_callable):
+        broker_or_callable()
+        return module, get_broker()
+
+    if not isinstance(broker_or_callable, Broker):
         raise ImportError("%r is not a Broker." % value)
-    return module, broker
+    return module, broker_or_callable
 
 
 def folder_path(value):
@@ -184,13 +191,18 @@ def make_argument_parser():
         "--fork-function", "-f", action="append", dest="forks", default=[],
         help="fork a subprocess to run the given function"
     )
+    parser.add_argument(
+        "--worker-shutdown-timeout", type=int, default=600000,
+        help="timeout for worker shutdown, in milliseconds (default: 10 minutes)"
+    )
 
     if HAS_WATCHDOG:
         parser.add_argument(
             "--watch", type=folder_path, metavar="DIR",
             help=(
                 "watch a directory and reload the workers when any source files "
-                "change (this feature must only be used during development)"
+                "change (this feature must only be used during development). "
+                "This option is currently only supproted on unix systems."
             )
         )
         parser.add_argument(
@@ -318,7 +330,7 @@ def watch_logs(log_filename, pipes, stop):
                 pipes = [p for p in pipes if not p.closed]
 
 
-def worker_process(args, worker_id, logging_pipe, canteen):
+def worker_process(args, worker_id, logging_pipe, canteen, event):
     try:
         # Re-seed the random number generator from urandom on
         # supported platforms.  This should make it so that worker
@@ -351,6 +363,11 @@ def worker_process(args, worker_id, logging_pipe, canteen):
     except ConnectionError:
         logger.exception("Broker connection failed.")
         return sys.exit(RET_CONNECT)
+    finally:
+        # Signal to the master process that this process has booted,
+        # regardless of whether it failed or not.  If it did fail, the
+        # worker process will realize that soon enough.
+        event.set()
 
     def termhandler(signum, frame):
         nonlocal running
@@ -373,8 +390,7 @@ def worker_process(args, worker_id, logging_pipe, canteen):
     while running:
         time.sleep(1)
 
-    timeout = int(os.getenv("dramatiq_worker_stop_timeout", '600000'))
-    worker.stop(timeout=timeout)
+    worker.stop(timeout=args.worker_shutdown_timeout)
     broker.close()
     logging_pipe.close()
 
@@ -437,16 +453,27 @@ def main(args=None):  # noqa
     canteen = multiprocessing.Value(Canteen)
     worker_pipes = []
     worker_processes = []
+    worker_process_events = []
     for worker_id in range(args.processes):
         read_pipe, write_pipe = multiprocessing.Pipe(duplex=False)
+        event = multiprocessing.Event()
         proc = multiprocessing.Process(
             target=worker_process,
-            args=(args, worker_id, StreamablePipe(write_pipe), canteen),
+            args=(args, worker_id, StreamablePipe(write_pipe), canteen, event),
             daemon=False,
         )
         proc.start()
         worker_pipes.append(read_pipe)
         worker_processes.append(proc)
+        worker_process_events.append(event)
+
+    # Wait for all worker processes to come online before starting the
+    # fork processes.  This is required to avoid race conditions like
+    # in #297.
+    for event in worker_process_events:
+        if proc.is_alive():
+            if not event.wait(timeout=30):
+                break
 
     fork_pipes = []
     fork_processes = []
@@ -481,6 +508,8 @@ def main(args=None):  # noqa
         )
 
     if HAS_WATCHDOG and args.watch:
+        if not hasattr(signal, "SIGHUP"):
+            raise RuntimeError("Watching for source changes is not supported on %s." % sys.platform)
         file_watcher = setup_file_watcher(args.watch, args.watch_use_polling)
 
     log_watcher_stop_event = Event()
@@ -547,7 +576,7 @@ def main(args=None):  # noqa
                 break
 
             else:
-                retcode = max(retcode, proc.exitcode)
+                retcode = retcode or proc.exitcode
 
     # The log watcher can't be a daemon in case we log to a file so we
     # have to wait for it to complete on exit.
@@ -563,4 +592,4 @@ def main(args=None):  # noqa
             return os.execvp(sys.executable, ["python", "-m", "dramatiq", *sys.argv[1:]])
         return os.execvp(sys.argv[0], sys.argv)
 
-    return retcode
+    return RET_KILLED if retcode < 0 else retcode