+++ dramatiq/worker.py
@@ -1,6 +1,6 @@
 # This file is a part of Dramatiq.
 #
-# Copyright (C) 2017,2018 CLEARTYPE SRL <bogdan@cleartype.io>
+# Copyright (C) 2017,2018,2019,2020 CLEARTYPE SRL <bogdan@cleartype.io>
 #
 # Dramatiq is free software; you can redistribute it and/or modify it
 # under the terms of the GNU Lesser General Public License as published by
@@ -482,8 +482,11 @@ class _WorkerThread(Thread):
             # to pass exceptions into results.
             message.stuff_exception(e)
 
+            throws = message.options.get("throws") or actor.options.get("throws")
             if isinstance(e, RateLimitExceeded):
                 self.logger.debug("Rate limit exceeded in message %s: %s.", message, e)
+            elif throws and isinstance(e, throws):
+                self.logger.info("Failed to process message %s with expected exception %s.", message, type(e).__name__)
             elif not isinstance(e, Retry):
                 self.logger.error("Failed to process message %s with unhandled exception.", message, exc_info=True)
 
@@ -496,6 +499,11 @@ class _WorkerThread(Thread):
             # this is safe.  Probably.
             self.consumers[message.queue_name].post_process_message(message)
             self.work_queue.task_done()
+
+            # See discussion #351.  Keeping a reference to the
+            # exception can lead to memory bloat because it may be a
+            # while before a GC triggers.
+            message.clear_exception()
 
     def pause(self):
         """Pause this worker.